+++ akka-http2-support/src/test/scala/akka/http/scaladsl/Http2BindingViaConfigSpec.scala
@@ -68,7 +68,7 @@ class Http2BindingViaConfigSpec extends AkkaSpec("""
       val ext: Http2Ext = Http2()
       // the Shadow is a structural type, so this assignment only works 
       // if all the methods it requires exist on the real class
-      val shadow: akka.http.scaladsl.Http2Shadow.ShadowHttp2Ext = ext
+      val shadow: akka.http.impl.engine.Http2Shadow.ShadowHttp2Ext = ext
     }
   }
